Cleaned up warnings Derived rides from CoreBase. Added RideName and StallName. Derived stall from CoreBase. Added Parent CoreBase aggregate class so path finder can locate them. Adjusted buildings. Fixed temp paths.
git-svn-id: https://tpb3dtg.svn.sourceforge.net/svnroot/tpb3dtg@40 7c74f4f4-39d4-4077-a54d-10deffb394f4

+++ AI/PeepAgent.cpp
@@ -52,7 +52,7 @@ void PeepsAgent::Update (float dt, int tod)
       Person* peep = Peeps->NewPerson();
       mPark.EnterPark (peep);
 
-      Peeps->Update( dt, tod);
+      Peeps->Update( dt, (float)tod);
    }
 
 //   if ( (rand() % 4) == 3 )  // TODO: need a better spawn mechanism, raised to 100
